// frontend/src/api/generated/models/dashboardHumanTaskDoc.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * acpms-server
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardHumanTaskDocAssignee } from './dashboardHumanTaskDocAssignee';

export interface DashboardHumanTaskDoc {
  /** @nullable */
  assignee?: DashboardHumanTaskDocAssignee;
  createdAt: string;
  description: string;
  id: string;
  projectId: string;
  projectName: string;
  title: string;
  type: string;
}